pyproject.toml
[project]
name = "funding-agent"
version = "0.1.0"
description = "Automated funding opportunity search agent for AI + transportation research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "jinja2>=3.1",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "python-dotenv>=1.0",
    "openai>=1.50",
    "certifi>=2024.0",
    "rich>=13.0",
    "tenacity>=8.2",
    # Web platform
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "aiosqlite>=0.20",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0,<4.1",
    "python-multipart>=0.0.20",
    "email-validator>=2.0",
    "PyMuPDF>=1.24",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
